Fix leak releasing child view controller. (#511)
When running with host application, rootViewController is retained by the window, and it retains the view controller we are testing (which is a child view controller). This causes a leak on the view controller we're testing.
By removing the view controller from it's parent, no one retains the view controller, fixing the leak.